refactor: cross-cutting modernization of syntax patterns, error handling, and API consistency
Comprehensive refinement pass across all franken_node subsystems (~116 files),
applying consistent code quality improvements without changing behavior:

Guard pattern modernization:
- Convert nested if-let patterns to Rust 1.80+ && guard chains across
  control_plane, connector, security, and runtime modules for reduced
  indentation and improved readability

CRDT Default trait implementation:
- Add explicit Default trait implementations for LwwMap, OrSet, GCounter,
  PnCounter and delegate new() constructors to Default::default() for
  Rust convention compliance

Method renaming (incident_bundle_retention):
- Rename from_str() to parse() on Severity, RetentionTier, ExportFormat
  to avoid confusion with FromStr trait convention

Error handling improvements:
- Replace unchecked unwrap() with expect() including context messages
  explaining why values are guaranteed present

Return expression simplification:
- Remove intermediate variable assignments in closure_proof() methods,
  directly returning call results

Feature addition (partner_lighthouse_programs):
- Add duplicate deployment and outcome detection with error codes
  PLP-ERR-003 and PLP-ERR-004, including validation tests preventing
  count drift

Miscellaneous:
- Add #[allow(dead_code)] attributes where needed (config.rs)
- Minor single-line consistency improvements across 20+ modules in
  connector, control_plane, security, observability, storage,
  supply_chain, and VEF subsystems

+++ crates/franken-node/src/policy/integrity_sweep_scheduler.rs
@@ -748,7 +748,7 @@ mod tests {
         for i in 1..=5 {
             let evidence = EvidenceTrajectory::new(
                 i * 2,
-                i as u32,
+                i,
                 0.9 - (i as f64 * 0.1),
                 Trend::Degrading,
                 i as u64,
